third_party/WebKit/LayoutTests/platform/win/external/wpt/pointerevents/pointerevent_lostpointercapture_is_first-manual-expected.txt

This is a testharness.js-based test.
PASS lostpointercapture is dispatched prior to subsequent events 
PASS PointerEvent Automation 
Harness: the test ran to completion.
